+++ lib/src/workflows/use_version.workflow.dart
@@ -4,6 +4,7 @@ import 'package:io/ansi.dart';
 
 import '../models/cache_flutter_version_model.dart';
 import '../models/project_model.dart';
+import '../services/cache_service.dart';
 import '../utils/helpers.dart';
 import '../utils/which.dart';
 import 'resolve_project_deps.workflow.dart';
@@ -28,6 +29,7 @@ class UseVersionWorkflow extends Workflow {
   }) async {
     if (!skipSetup) {
       await get<SetupFlutterWorkflow>()(version);
+      version = get<CacheService>().getVersion(version) ?? version;
     }
 
     get<VerifyProjectWorkflow>()(project, force: force);